refactor(net): rename nclutils.network to nclutils.net
