+++ src/nclutils/net/__init__.py
@@ -1,5 +1,5 @@
 """Network utilities."""
 
-from .network import network_available
+from .net import network_available
 
 __all__ = ["network_available"]
